Moved some more fixtures into conftest
